+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/domestic/scheduled/payments/consents/api/v3_1_8/CreateDomesticScheduledPaymentsConsents.kt
@@ -1,0 +1,199 @@
+package com.forgerock.uk.openbanking.tests.functional.payment.domestic.scheduled.payments.consents.api.v3_1_8
+
+import assertk.assertThat
+import assertk.assertions.contains
+import assertk.assertions.isEqualTo
+import assertk.assertions.isNotEmpty
+import assertk.assertions.isNotNull
+import com.forgerock.securebanking.framework.conditions.Status
+import com.forgerock.securebanking.framework.configuration.psu
+import com.forgerock.securebanking.framework.data.AccessToken
+import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
+import com.forgerock.securebanking.framework.http.fuel.defaultMapper
+import com.forgerock.securebanking.framework.signature.signPayloadSubmitPayment
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
+import com.forgerock.uk.openbanking.framework.constants.INVALID_FORMAT_DETACHED_JWS
+import com.forgerock.uk.openbanking.framework.constants.INVALID_SIGNING_KID
+import com.forgerock.uk.openbanking.framework.errors.INVALID_FORMAT_DETACHED_JWS_ERROR
+import com.forgerock.uk.openbanking.framework.errors.NO_DETACHED_JWS
+import com.forgerock.uk.openbanking.framework.errors.REQUEST_EXECUTION_TIME_IN_THE_PAST
+import com.forgerock.uk.openbanking.framework.errors.UNAUTHORIZED
+import com.forgerock.uk.openbanking.support.discovery.getPaymentsApiLinks
+import com.forgerock.uk.openbanking.support.payment.PaymentAS
+import com.forgerock.uk.openbanking.support.payment.PaymentRS
+import com.github.kittinunf.fuel.core.FuelError
+import org.assertj.core.api.Assertions
+import org.joda.time.DateTime
+import uk.org.openbanking.datamodel.payment.OBWriteDomesticScheduledConsent4
+import uk.org.openbanking.datamodel.payment.OBWriteDomesticScheduledConsentResponse5
+import uk.org.openbanking.testsupport.payment.OBWriteDomesticScheduledConsentTestDataFactory
+
+class CreateDomesticScheduledPaymentsConsents(val version: OBVersion, val tppResource: CreateTppCallback.TppResource) {
+
+    val paymentLinks = getPaymentsApiLinks(version)
+
+    fun createDomesticScheduledPaymentsConsentsTest() {
+        // Given
+        val consentRequest = OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4()
+        val consent = createDomesticScheduledPaymentConsent(consentRequest)
+
+        // Then
+        assertThat(consent).isNotNull()
+        assertThat(consent.data).isNotNull()
+        assertThat(consent.data.consentId).isNotEmpty()
+        Assertions.assertThat(consent.data.status.toString()).`is`(Status.consentCondition)
+        assertThat(consent.risk).isNotNull()
+    }
+
+    fun shouldCreateDomesticScheduledPaymentsConsents_throwsSendInvalidFormatDetachedJwsTest() {
+        // Given
+        val consentRequest = OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4()
+
+        // When
+        val exception = org.junit.jupiter.api.Assertions.assertThrows(AssertionError::class.java) {
+            PaymentRS().consent<OBWriteDomesticScheduledConsentResponse5>(
+                paymentLinks.CreateDomesticScheduledPaymentConsent,
+                consentRequest,
+                tppResource.tpp,
+                version,
+                INVALID_FORMAT_DETACHED_JWS
+            )
+        }
+
+        // Then
+        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(400)
+        assertThat(exception.message.toString()).contains(INVALID_FORMAT_DETACHED_JWS_ERROR)
+    }
+
+    fun shouldCreateDomesticScheduledPaymentsConsents_throwsNoDetachedJwsTest() {
+        // Given
+        val consentRequest = OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4()
+
+        // When
+        val exception = org.junit.jupiter.api.Assertions.assertThrows(AssertionError::class.java) {
+            PaymentRS().consentNoDetachedJwt<OBWriteDomesticScheduledConsentResponse5>(
+                paymentLinks.CreateDomesticScheduledPaymentConsent,
+                consentRequest,
+                tppResource.tpp,
+                version
+            )
+        }
+
+        // Then
+        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(400)
+        assertThat(exception.message.toString()).contains(NO_DETACHED_JWS)
+    }
+
+    fun shouldCreateDomesticScheduledPaymentsConsents_throwsNotPermittedB64HeaderAddedInTheDetachedJwsTest() {
+        // Given
+        val consentRequest = OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4()
+
+        val signedPayload =
+            signPayloadSubmitPayment(
+                defaultMapper.writeValueAsString(consentRequest),
+                tppResource.tpp.signingKey,
+                tppResource.tpp.signingKid,
+                true
+            )
+
+        // When
+        val exception = org.junit.jupiter.api.Assertions.assertThrows(AssertionError::class.java) {
+            PaymentRS().consent<OBWriteDomesticScheduledConsentResponse5>(
+                paymentLinks.CreateDomesticScheduledPaymentConsent,
+                consentRequest,
+                tppResource.tpp,
+                version,
+                signedPayload
+            )
+        }
+
+        // Then
+        assertThat((exception.cause as FuelError).response.responseMessage).isEqualTo(UNAUTHORIZED)
+        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(401)
+    }
+
+    fun shouldCreateDomesticScheduledPaymentsConsents_throwsSendInvalidKidDetachedJwsTest() {
+        // Given
+        val consentRequest = OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4()
+
+        val signedPayloadConsent =
+            signPayloadSubmitPayment(
+                defaultMapper.writeValueAsString(consentRequest),
+                tppResource.tpp.signingKey,
+                INVALID_SIGNING_KID
+            )
+
+        // When
+        val exception = org.junit.jupiter.api.Assertions.assertThrows(AssertionError::class.java) {
+            PaymentRS().consent<OBWriteDomesticScheduledConsentResponse5>(
+                paymentLinks.CreateDomesticScheduledPaymentConsent,
+                consentRequest,
+                tppResource.tpp,
+                version,
+                signedPayloadConsent
+            )
+        }
+
+        // Then
+        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(401)
+        assertThat((exception.cause as FuelError).response.responseMessage).isEqualTo(UNAUTHORIZED)
+    }
+
+    fun shouldCreateDomesticScheduledPaymentsConsents_throwsRequestExecutionTimeInThePastTest() {
+        // Given
+        val consentRequest = OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4()
+        consentRequest.data.initiation.requestedExecutionDateTime = DateTime.now().minusDays(1)
+
+        val signedPayloadConsent =
+            signPayloadSubmitPayment(
+                defaultMapper.writeValueAsString(consentRequest),
+                tppResource.tpp.signingKey,
+                INVALID_SIGNING_KID
+            )
+
+        // When
+        val exception = org.junit.jupiter.api.Assertions.assertThrows(AssertionError::class.java) {
+            PaymentRS().consent<OBWriteDomesticScheduledConsentResponse5>(
+                paymentLinks.CreateDomesticScheduledPaymentConsent,
+                consentRequest,
+                tppResource.tpp,
+                version,
+                signedPayloadConsent
+            )
+        }
+
+        // Then
+        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(400)
+        assertThat(exception.message.toString()).contains(REQUEST_EXECUTION_TIME_IN_THE_PAST)
+    }
+
+    fun createDomesticScheduledPaymentConsent(consentRequest: OBWriteDomesticScheduledConsent4): OBWriteDomesticScheduledConsentResponse5 {
+        val signedPayloadConsent =
+            signPayloadSubmitPayment(
+                defaultMapper.writeValueAsString(consentRequest),
+                tppResource.tpp.signingKey,
+                tppResource.tpp.signingKid
+            )
+
+        // When
+        val consent = PaymentRS().consent<OBWriteDomesticScheduledConsentResponse5>(
+            paymentLinks.CreateDomesticScheduledPaymentConsent,
+            consentRequest,
+            tppResource.tpp,
+            version,
+            signedPayloadConsent
+        )
+        return consent
+    }
+
+    fun createDomesticScheduledPaymentConsentAndGetAccessToken(consentRequest: OBWriteDomesticScheduledConsent4): Pair<OBWriteDomesticScheduledConsentResponse5, AccessToken> {
+        val consent = createDomesticScheduledPaymentConsent(consentRequest)
+        val accessTokenAuthorizationCode = PaymentAS().getAccessToken(
+            consent.data.consentId,
+            tppResource.tpp.registrationResponse,
+            psu,
+            tppResource.tpp
+        )
+        return consent to accessTokenAuthorizationCode
+    }
+}
